pyproject.toml
[tool.poetry]
name = "hierarchical-data-and-you-gmferise"
version = "0.1.0"
description = ""
authors = ["Gavin Ferise <gmferise@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
django-mptt = "^0.13.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
